refactor(register.post.ts): implementa validações
